model_utils.py
import torch
import torch.nn as nn
import timm
import numpy as np
from PIL import Image
import torchvision.transforms as transforms

class EfficientNetDetector(nn.Module):
    def __init__(self, num_classes=2, dropout=0.3):
        super(EfficientNetDetector, self).__init__()
        self.backbone = timm.create_model('efficientnet_b3', pretrained=True)
        num_features = self.backbone.classifier.in_features
        self.backbone.classifier = nn.Sequential(
            nn.Dropout(dropout),
            nn.Linear(num_features, 512),
            nn.ReLU(),
            nn.Dropout(dropout/2),
            nn.Linear(512, num_classes)
        )
    
    def forward(self, x):
        return self.backbone(x)

class VisionTransformerDetector(nn.Module):
    def __init__(self, num_classes=2, dropout=0.3):
        super(VisionTransformerDetector, self).__init__()
        self.backbone = timm.create_model('vit_base_patch16_224', pretrained=True)
        num_features = self.backbone.head.in_features
        self.backbone.head = nn.Sequential(
            nn.Dropout(dropout),
            nn.Linear(num_features, 512),
            nn.ReLU(),
            nn.Dropout(dropout/2),
            nn.Linear(512, num_classes)
        )
    
    def forward(self, x):
        return self.backbone(x)

class XceptionDetector(nn.Module):
    def __init__(self, num_classes=2, dropout=0.3):
        super(XceptionDetector, self).__init__()
        self.backbone = timm.create_model('xception', pretrained=True)
        num_features = self.backbone.fc.in_features
        self.backbone.fc = nn.Sequential(
            nn.Dropout(dropout),
            nn.Linear(num_features, 512),
            nn.ReLU(),
            nn.Dropout(dropout/2),
            nn.Linear(512, num_classes)
        )
    
    def forward(self, x):
        return self.backbone(x)

def get_image_transform():
    """Get the standard image transformation pipeline"""
    return transforms.Compose([
        transforms.Resize((224, 224)),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
    ])

def load_model_weights(model, model_path, device):
    """Load model weights from checkpoint"""
    try:
        model.load_state_dict(torch.load(model_path, map_location=device))
        model.eval()
        model.to(device)
        return True
    except Exception as e:
        print(f"Error loading model weights: {e}")
        return False

def predict_single_image(model, image, device, transform):
    """
    Make prediction on a single image
    
    Args:
        model: PyTorch model
        image: PIL Image
        device: torch device
        transform: torchvision transforms
    
    Returns:
        predicted_class (int): 0 for real, 1 for fake
        confidence (float): confidence score
        probabilities (np.array): class probabilities
    """
    try:
        # Ensure RGB format
        if image.mode != 'RGB':
            image = image.convert('RGB')
        
        # Apply transforms and add batch dimension
        image_tensor = transform(image).unsqueeze(0).to(device)
        
        # Make prediction
        with torch.no_grad():
            outputs = model(image_tensor)
            probabilities = torch.softmax(outputs, dim=1)
            predicted_class = torch.argmax(probabilities, dim=1).item()
            confidence = probabilities[0][predicted_class].item()
        
        return predicted_class, confidence, probabilities[0].cpu().numpy()
    
    except Exception as e:
        print(f"Error during prediction: {e}")
        return None, None, None

def batch_predict_images(model, images, device, transform, progress_callback=None):
    """
    Make predictions on multiple images
    
    Args:
        model: PyTorch model
        images: List of PIL Images
        device: torch device
        transform: torchvision transforms
        progress_callback: Optional callback function for progress updates
    
    Returns:
        List of prediction results
    """
    results = []
    
    for i, image in enumerate(images):
        predicted_class, confidence, probabilities = predict_single_image(
            model, image, device, transform
        )
        
        if predicted_class is not None:
            results.append({
                'prediction': predicted_class,
                'confidence': confidence,
                'probabilities': probabilities
            })
        else:
            results.append({
                'prediction': None,
                'confidence': None,
                'probabilities': None
            })
        
        # Call progress callback if provided
        if progress_callback:
            progress_callback(i + 1, len(images))
    
    return results

def get_model_info():
    """Get information about available models"""
    return {
        'Vision Transformer': {
            'class': VisionTransformerDetector,
            'accuracy': 99.92,
            'description': 'State-of-the-art Vision Transformer with attention mechanism',
            'parameters': '86.6M',
            'inference_time': '45ms'
        },
        'Xception': {
            'class': XceptionDetector,
            'accuracy': 99.26,
            'description': 'Efficient Xception architecture with depthwise separable convolutions',
            'parameters': '22.9M',
            'inference_time': '23ms'
        },
        'EfficientNet B3': {
            'class': EfficientNetDetector,
            'accuracy': 99.11,
            'description': 'Balanced efficiency and performance with compound scaling',
            'parameters': '12.2M',
            'inference_time': '18ms'
        }
    }

def create_model_summary():
    """Create a summary of model performance metrics"""
    return {
        'Vision Transformer': {
            'accuracy': 99.92,
            'precision': 99.92,
            'recall': 99.92,
            'f1_score': 99.92,
            'true_positives': 4996,
            'true_negatives': 4996,
            'false_positives': 4,
            'false_negatives': 4
        },
        'Xception': {
            'accuracy': 99.26,
            'precision': 99.25,
            'recall': 99.27,
            'f1_score': 99.26,
            'true_positives': 4963,
            'true_negatives': 4967,
            'false_positives': 37,
            'false_negatives': 33
        },
        'EfficientNet B3': {
            'accuracy': 99.11,
            'precision': 99.10,
            'recall': 99.12,
            'f1_score': 99.11,
            'true_positives': 4956,
            'true_negatives': 4955,
            'false_positives': 44,
            'false_negatives': 45
        }
    } 