requirements.txt
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
timm==0.9.7
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
plotly==5.15.0
albumentations==1.3.1
opencv-python-headless==4.8.0.76
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2 